+++ web/public/assets/js/app/node-page-charts/display-formatters.js
@@ -1,0 +1,222 @@
+/*
+ * Copyright © 2025-26 l5yth & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Display-only formatters used by the node detail page outside the chart
+ * SVG rendering path (hardware model, coordinates, message timestamps,
+ * uptimes, batteries, etc.).  Sibling to ``format-utils.js`` so chart code
+ * only carries chart concerns.
+ *
+ * @module node-page-charts/display-formatters
+ */
+
+import { numberOrNull, stringOrNull } from '../value-helpers.js';
+import { padTwo } from './format-utils.js';
+
+/**
+ * Format a frequency value using MHz units when a numeric reading is
+ * available.  Non-numeric input is passed through unchanged.
+ *
+ * @param {*} value Raw frequency value.
+ * @returns {string|null} Formatted frequency string or ``null``.
+ */
+export function formatFrequency(value) {
+  if (value == null || value === '') return null;
+  const numeric = numberOrNull(value);
+  if (numeric == null) {
+    return stringOrNull(value);
+  }
+  const abs = Math.abs(numeric);
+  if (abs >= 1_000_000) {
+    return `${(numeric / 1_000_000).toFixed(3)} MHz`;
+  }
+  if (abs >= 1_000) {
+    return `${(numeric / 1_000).toFixed(3)} MHz`;
+  }
+  return `${numeric.toFixed(3)} MHz`;
+}
+
+/**
+ * Format a battery reading as a percentage with one decimal place.
+ *
+ * @param {*} value Raw battery value.
+ * @returns {string|null} Formatted percentage or ``null``.
+ */
+export function formatBattery(value) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return null;
+  return `${numeric.toFixed(1)}%`;
+}
+
+/**
+ * Format a voltage reading with two decimal places.
+ *
+ * @param {*} value Raw voltage value.
+ * @returns {string|null} Formatted voltage string or ``null``.
+ */
+export function formatVoltage(value) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return null;
+  return `${numeric.toFixed(2)} V`;
+}
+
+/**
+ * Convert an uptime reading in seconds to a concise human-readable string.
+ *
+ * @param {*} value Raw uptime value.
+ * @returns {string|null} Formatted uptime string or ``null`` when invalid.
+ */
+export function formatUptime(value) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return null;
+  const seconds = Math.floor(numeric);
+  const parts = [];
+  const days = Math.floor(seconds / 86_400);
+  if (days > 0) parts.push(`${days}d`);
+  const hours = Math.floor((seconds % 86_400) / 3_600);
+  if (hours > 0) parts.push(`${hours}h`);
+  const minutes = Math.floor((seconds % 3_600) / 60);
+  if (minutes > 0) parts.push(`${minutes}m`);
+  const remainSeconds = seconds % 60;
+  if (parts.length === 0 || remainSeconds > 0) {
+    parts.push(`${remainSeconds}s`);
+  }
+  return parts.join(' ');
+}
+
+/**
+ * Format a timestamp for the message log as ``YYYY-MM-DD HH:MM`` in the
+ * local time zone.
+ *
+ * @param {*} value Seconds since the epoch.
+ * @param {string|null} [isoFallback] ISO timestamp to prefer when available.
+ * @returns {string|null} Formatted timestamp string or ``null``.
+ */
+export function formatMessageTimestamp(value, isoFallback = null) {
+  const iso = stringOrNull(isoFallback);
+  let date = null;
+  if (iso) {
+    const candidate = new Date(iso);
+    if (!Number.isNaN(candidate.getTime())) {
+      date = candidate;
+    }
+  }
+  if (!date) {
+    const numeric = numberOrNull(value);
+    if (numeric == null) return null;
+    const candidate = new Date(numeric * 1000);
+    if (Number.isNaN(candidate.getTime())) {
+      return null;
+    }
+    date = candidate;
+  }
+  const year = date.getFullYear();
+  const month = padTwo(date.getMonth() + 1);
+  const day = padTwo(date.getDate());
+  const hours = padTwo(date.getHours());
+  const minutes = padTwo(date.getMinutes());
+  return `${year}-${month}-${day} ${hours}:${minutes}`;
+}
+
+/**
+ * Format a hardware model string while hiding unset placeholders.
+ *
+ * Firmware uses the literal string ``"UNSET"`` for nodes that have not
+ * reported a hardware model; this helper suppresses that value so the UI
+ * displays an empty cell instead.
+ *
+ * @param {*} value Raw hardware model value.
+ * @returns {string} Sanitised hardware model string, or empty string.
+ */
+export function formatHardwareModel(value) {
+  const text = stringOrNull(value);
+  if (!text || text.toUpperCase() === 'UNSET') {
+    return '';
+  }
+  return text;
+}
+
+/**
+ * Format a geographic coordinate with consistent decimal precision.
+ *
+ * @param {*} value Raw coordinate value.
+ * @param {number} [precision=5] Number of decimal places.
+ * @returns {string} Formatted coordinate string, or empty string when invalid.
+ */
+export function formatCoordinate(value, precision = 5) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return '';
+  return numeric.toFixed(precision);
+}
+
+/**
+ * Convert an absolute UNIX timestamp into a relative time description.
+ *
+ * Returns strings such as ``"42s"``, ``"3m 15s"``, ``"2h 5m"``, ``"1d 3h"``.
+ *
+ * @param {*} value Raw timestamp expressed in seconds since the epoch.
+ * @param {number} [referenceSeconds] Optional reference timestamp in seconds.
+ *   Defaults to ``Date.now() / 1000``.
+ * @returns {string} Relative time string or empty string when unavailable.
+ */
+export function formatRelativeSeconds(value, referenceSeconds = Date.now() / 1000) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return '';
+  const reference = numberOrNull(referenceSeconds);
+  const base = reference != null ? reference : Date.now() / 1000;
+  const diff = Math.floor(base - numeric);
+  const safeDiff = Number.isFinite(diff) ? Math.max(diff, 0) : 0;
+  if (safeDiff < 60) return `${safeDiff}s`;
+  if (safeDiff < 3_600) {
+    const minutes = Math.floor(safeDiff / 60);
+    const seconds = safeDiff % 60;
+    return seconds > 0 ? `${minutes}m ${seconds}s` : `${minutes}m`;
+  }
+  if (safeDiff < 86_400) {
+    const hours = Math.floor(safeDiff / 3_600);
+    const minutes = Math.floor((safeDiff % 3_600) / 60);
+    return minutes > 0 ? `${hours}h ${minutes}m` : `${hours}h`;
+  }
+  const days = Math.floor(safeDiff / 86_400);
+  const hours = Math.floor((safeDiff % 86_400) / 3_600);
+  return hours > 0 ? `${days}d ${hours}h` : `${days}d`;
+}
+
+/**
+ * Format a duration expressed in seconds using a compact human-readable form.
+ *
+ * @param {*} value Raw duration in seconds.
+ * @returns {string} Human-readable duration string or empty string.
+ */
+export function formatDurationSeconds(value) {
+  const numeric = numberOrNull(value);
+  if (numeric == null) return '';
+  const duration = Math.max(Math.floor(numeric), 0);
+  if (duration < 60) return `${duration}s`;
+  if (duration < 3_600) {
+    const minutes = Math.floor(duration / 60);
+    const seconds = duration % 60;
+    return seconds > 0 ? `${minutes}m ${seconds}s` : `${minutes}m`;
+  }
+  if (duration < 86_400) {
+    const hours = Math.floor(duration / 3_600);
+    const minutes = Math.floor((duration % 3_600) / 60);
+    return minutes > 0 ? `${hours}h ${minutes}m` : `${hours}h`;
+  }
+  const days = Math.floor(duration / 86_400);
+  const hours = Math.floor((duration % 86_400) / 3_600);
+  return hours > 0 ? `${days}d ${hours}h` : `${days}d`;
+}
